requirements.txt
Flask-Uploads
Flask-Login
Flask-Dotenv
Werkzeug
Flask==3.0.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.3
Gunicorn
